ThreeStuckVars and NumberDivider tasks commit after the first testing

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -2,6 +2,6 @@ package lang.print.gaps.task5;
 
 public class NumberDivider {
     void divide(int toBeDivided) {
-        System.out.println(toBeDivided/5);
+        System.out.println((double) toBeDivided/5);
     }
 }
